Create vendorCheckN
You can use a convert switch instead of just an if statement to clean the code up, I am just learning about convert switches. By the end of the build of this project it needs to store vendor check in times in an SQL table and retrieve them when called to print. "//" starts my comment lines and each one describes what the code is for.